pyproject.toml
[project]
name = "twelve"
version = "0.1.0"
description = "Static sites with an extra 'push'."
readme = "README.md"
authors = [
    { name = "stephen gilmore"}
]
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "pymdown-extensions>=10.21.2",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.4",
    "rich>=15.0.0",
    "livereload>=2.7.1",
    "pyvips[binary]>=3.1.1",
    "rich-argparse>=1.7.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.10",
    "fake-useragent>=2.2.0",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-randomly>=4.1.0",
    
]

[project.scripts]
twelve = "twelve.cli:cli"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"


[tool.ruff]
line-length = 90
cache-dir = ".tmp/.ruff_cache"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --cov-report=xml:.tmp/coverage.xml"
testpaths = ["tests"]
cache_dir = ".tmp/.pytest_cache"

[tool.coverage.run]
branch = true
data_file = ".tmp/.coverage"
source = ["src"]
omit = []

[tool.coverage.report]
skip_covered = false
skip_empty = true
fail_under = 90
